backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
opencv-python
PyWavelets
numpy
scipy
reportlab